+++ R/TaskSurv.R
@@ -49,31 +49,28 @@ TaskSurv = R6::R6Class("TaskSurv",
     #' @param type (`character(1)`)\cr
     #' Name of the column giving the type of censoring. Default is 'right' censoring.
     initialize = function(id, backend, time, event, time2,
-      type = c("right", "left", "counting", "interval", "interval2", "mstate")) {
+      type = c("right", "left", "counting", "interval", "mstate")) {
+
       type = match.arg(type)
+
+      event = backend[, "event"][[1L]]
+      if (type == "mstate") {
+        assert_factor(event)
+      } else if (!is.logical(event)) {
+        assert_integerish(event, lower = 0, upper = 2)
+      }
+
+      private$.censtype = type
 
       if (type %in% c("right", "left", "mstate")) {
         super$initialize(
           id = id, task_type = "surv", backend = backend,
           target = c(time, event))
-      } else if (type %in% "interval2") {
-        super$initialize(
-          id = id, task_type = "surv", backend = backend,
-          target = c(time, time2))
       } else {
         super$initialize(
           id = id, task_type = "surv", backend = backend,
           target = c(time, time2, event))
       }
-
-      if (type %nin% "interval2") {
-        event = self$data(cols = event)[[1L]]
-        if (!is.logical(event)) {
-          assert_integerish(event, lower = 0, upper = 3)
-        }
-      }
-
-      private$.censtype = type
     },
 
     #' @description
@@ -85,9 +82,7 @@ TaskSurv = R6::R6Class("TaskSurv",
       tn = self$target_names
       d = self$data(rows, cols = self$target_names)
       args = list(time = d[[tn[1L]]], type = self$censtype)
-      if (self$censtype %in% "interval2") {
-        args$time2 = d[[tn[2L]]]
-      } else if (length(tn) == 2) {
+      if (length(tn) == 2) {
         args$event = as.integer(d[[tn[2L]]])
       } else {
         args$event = as.integer(d[[tn[3L]]])
@@ -111,9 +106,7 @@ TaskSurv = R6::R6Class("TaskSurv",
     formula = function(rhs = NULL) {
       # formula appends the rhs argument to Surv(time, event)~
       tn = self$target_names
-      if (self$censtype %in% "interval2") {
-        lhs = sprintf("Surv(time = %s, time2 = %s, type = 'interval2')", tn[1L], tn[2L])
-      } else if (length(tn) == 2) {
+      if (length(tn) == 2) {
         lhs = sprintf("Surv(%s, %s, type = '%s')", tn[1L], tn[2L], self$censtype)
       } else {
         lhs = sprintf("Surv(%s, %s, %s, type = '%s')", tn[1L], tn[2L], tn[3L], self$censtype)
@@ -124,10 +117,55 @@ TaskSurv = R6::R6Class("TaskSurv",
 
   active = list(
     #' @field censtype `character(1)`\cr
-    #' Returns the type of censoring, one of "right", "left", "counting", "interval", "interval2"
-    #' or "mstate".
+    #' Returns the type of censoring, one of "right", "left", "counting", "interval", or "mstate".
     censtype = function() {
       return(private$.censtype)
+    },
+
+    #' @field times `numeric()`\cr
+    #' Returns the (unsorted) true outcome times.
+    times = function(rows = NULL) {
+      truth = self$truth(rows)
+      if (censtype %in% c("interval", "counting")) {
+        return(truth[, 1:2])
+      } else {
+        return(truth[, 1])
+      }
+    },
+
+    #' @field status (`integer()`) \cr
+    #' Returns the true event indicator.
+    status = function(rows = NULL) {
+      truth = self$truth(rows)
+      if (censtype %in% c("interval", "counting")) {
+        status = truth[, 3L]
+      } else {
+        status = truth[, 2L]
+      }
+
+      as.integer(status)
+    },
+
+    #' @field unique_times (`numeric()`)\cr
+    #' Returns the sorted unique outcome times.
+    unique_times = function(rows = NULL) {
+      sort(unique(self$times(rows)))
+    },
+
+    #' @field unique_event_times (`numeric()`)\cr
+    #' Returns the sorted unique event (or failure) times.
+    unique_event_times = function(rows = NULL) {
+      sort(unique(self$times(rows)[self$status(rows) != 0]))
+    },
+
+    #' @field risk_set (`integer()`) \cr
+    #' Returns the `row_ids` of the observations 'at risk' (not dead or censored) just before `time`.
+    risk_set = function(time = NULL) {
+      if (is.null(time)) {
+        self$row_ids
+      } else {
+        self$row_ids[self$times() >= time]
+      }
     }
   ),